# api/app_management/openapi.yaml
openapi: 3.0.3
info:
  title: CasaOS App Management API
  version: v2
  description: |
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_dark_night_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
        <img alt="CasaOS" src="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
    </picture>

    CasaOS App Management manages app lifecycle, including installing, uninstalling, starting, stopping, setting, etc.

    For issues and discussions, please visit the [GitHub repository](https://github.com/IceWhaleTech/CasaOS) or join [our Discord](https://discord.gg/knqAbbBbeX).
servers:
  - url: /v2/app_management
...
security:
  - access_token: []
paths:
...
  /updates:
    get:
      summary: List installed apps with update checks, operation status and rollback availability
      operationId: appUpdates
      tags: [Compose methods]
      responses:
        "200":
          description: Installed app update status; contains no saved configuration or credentials
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppUpdateStatus"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /updates/check:
    post:
      summary: Refresh store catalogs and check every installed app without pulling images
      operationId: checkAppUpdates
      tags: [Compose methods]
      responses:
        "200":
          description: Refreshed statuses, including per-app check failures
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppUpdateStatus"
        "409":
          $ref: "#/components/responses/ResponseConflict"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /compose/{id}/rollback:
    post:
      summary: Recreate an app using its saved previous images and configuration, keeping current volumes
      operationId: rollbackComposeApp
      tags: [Compose methods]
      parameters:
        - $ref: "#/components/parameters/ComposeAppID"
      responses:
        "202":
          description: Rollback started; poll GET /updates for its result
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "409":
          $ref: "#/components/responses/ResponseConflict"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
  parameters:
...
    ComposeAppID:
      name: id
      description: App ID of a compose app (equivalent to compose project name)
      in: path
      required: true
      schema:
        type: string
        example: "syncthing"
...
  responses:
...
    ResponseInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Internal Server Error"
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Not Found"
...
    ResponseConflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Conflict"
...
  schemas:
    AppUpdateStatus:
      type: object
      required: [id, title, icon, current_version, check_status, operation, rollback_available]
      properties:
        id:
          type: string
        title:
          type: object
          additionalProperties:
            type: string
        icon:
          type: string
        current_version:
          type: string
        target_version:
          type: string
        check_status:
          type: string
          enum: [unchecked, available, up_to_date, unmanaged, failed]
        checked_at:
          type: string
          format: date-time
        check_error:
          type: string
        operation:
          type: string
          enum: [idle, busy, preparing, pulling, applying, reverting, updated, reverted, failed, interrupted]
        error:
          type: string
        rollback_available:
          type: boolean
        rollback_version:
          type: string
        rollback_date:
          type: string
          format: date-time
        rollback_reason:
          type: string
    BaseResponse:
      properties:
        message:
          description: message returned by server side if there is any
          type: string
          example: ""
...